game: Make medic's passive heal deterministic
Medic's currently heal every second via 'ClientTimerActions', which
loops continuously in the background.

This means that in case a Medic takes damage, the timer could be
anywhere between 0.1 and 0.9 seconds off the next passive heal,
effectively making it random in a way that neither the player nor the
opponent can control.

This commit alleviates the issue by resetting and pausing the timer once
the player has reached MAX_HEALTH, making the passive heal
deterministic.