etmain: Tweak 3P 'jump_nostep_1h'
The 'jump_nostep_1h' jump animation does not only animate jumping off
the ground, but also include animations for landing on the ground again.

This has the following downsides:

- If the jump distance is longer, then the animation plays the included
  "landing on the ground" part mid-air.

- As the jump animations take priority over anything else, if the
  distance is too small, the player will "hover"/"glide" as the "landing
  on the ground" part is still playing.

In practice this means that this animation causes a lot of
"hovering"/"gliding".

This is addressed by simply trimming the "landing on the ground" part
from it.