etmain: Canonicalize 3P 'aimless' weapon animations
This commit introduces a new weapons category 'aimless' that subsumes
all weapons that should not be held in front of the player as if they
could aim with them.

Along with this refactor, knives and pliers (named 'special') are both
now also held in this way, instead of holding them like pistols.