game: Support anims playing only while firing
When a character animation sets 'loop' to '-1', the animation will now
loop over it's whole frameNum duration AND be cancelled by other
animations once the player stops firing.

'Pliers' of engi is a classic use-case for this, where a single plier
"attack" would cause the whole duration of the character plier firing
animation being played, which is much longer.

This addition allows such cases to stop when the player stops firing,
thus only playing on continuous use.