etmain: Tweak airplane visuals
Modified textures with added pseudo-shadows and some grit.

Also removed the transparency effects, as the models caused transparency
overlap between their own faces and there doesn't seem to be a need for
it atm. - it however should be fixed and re-enabled at a later point,
but this is not possible without recasting the model geom and doing the
UV wrap again.