cgame: Add ETL dynamic gun fov
Instead of having only 1 fixed 'cg_gunFov' for all guns for all possible
values of 'cg_fov', a new dynamic mode is added to set the hand axis
when 'cg_gunFov' is set to '0', which is now the default.

For each gun, 2 sets of hand axis modifier references were determined,
one that used 'cg_fov 90' as a reference and a separate set that used
'cg_fov 120' as reference.
(The process by which these modifiers were determined is documented in
code)

This way it is ensured that for cg_fov 90 and 120 we ship predictable
results that ensure visibility, consistency and glitch-freeness.

Other 'cg_fov' values are then successfully linearly interpolated among
the difference of these 2 sets, resulting in consistent results for any
cg_fov between 90 to 120 and best effort results for beyond or below.